+++ src/main/java/com/poimapper/config/ExcelRowBeanMapperOptions.java
@@ -14,7 +14,7 @@ public class ExcelRowBeanMapperOptions {
 
     private String datePattern = CastConstants.DATE_PATTERN;
     private String dateTimePattern = CastConstants.DATE_TIME_PATTERN;
-    private Map<String,Object> mapperSettings = Map.of("strictMode",false);
+    private PoiBuilderConfig poiBuilderConfig;
     private CastString customCastingFunc;
-    private LinkedHashMap<String, Map<String,Object>> rowMappingOptions;
+    private LinkedHashMap<String, PoiConfig> rowMappingOptions;
 }