+++ src/main/java/com/poimapper/config/PoiBuilderConfig.java
@@ -1,0 +1,10 @@
+package com.poimapper.config;
+
+import lombok.Data;
+
+@Data
+public class PoiBuilderConfig {
+
+    Boolean suppressWarnings;
+    Boolean strictMode;
+}