+++ src/test/java/UserInfo.java
@@ -1,7 +1,5 @@
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.RequiredArgsConstructor;
+import lombok.*;
+
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -10,6 +8,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @RequiredArgsConstructor
 @EqualsAndHashCode
+@Builder
 public class UserInfo {
     private String name;
     private LocalDate birthDate;